Making task ID 32 bit for Linux

+++ nutmeglib/src/server/task.h
@@ -110,7 +110,7 @@ public:
     QVariantList args;
 
     QUuid sessionId;
-    uint64_t taskId;
+    uint32_t taskId;
 
 };
